scraper/requirements.txt
requests==2.32.3
PyYAML==6.0.2
feedparser==6.0.11
python-dotenv==1.0.1
